Variant tables , paging is broken Improve pathing
