+++ src/index.js
@@ -10,15 +10,28 @@ import ListDatasets from './ListDatasets.js'
 import ListReferenceSets from './ListReferenceSets.js'
 import ReferenceSet from './ReferenceSet.js'
 import ListReferences from './ListReferences.js'
+import VariantSet from './VariantSet.js'
+import ReadGroupSet from './ReadGroupSet.js'
+import FeatureSet from './FeatureSet.js'
+import SearchVariants from './SearchVariants.js'
+import ListVariantSets from './ListVariantSets.js'
+
+let baseurl = "http://1kgenomes.ga4gh.org/"
 
 ReactDOM.render(<Router history={browserHistory}>
     <Route path="/" component={App}>
-      <Route path="datasets" baseurl="http://1kgenomes.ga4gh.org/" component={ListDatasets} >
-        <Route path=":id" baseurl="http://1kgenomes.ga4gh.org/" component={Dataset} />
+      <Route path="datasets" baseurl={baseurl} component={ListDatasets}/>
+      <Route path="datasets/:datasetId" baseurl={baseurl} component={Dataset}>
+        <Route path="variantsets" baseurl={baseurl} component={ListVariantSets}/>
       </Route>
-      <Route path="referencesets" baseurl="http://1kgenomes.ga4gh.org/" component={ListReferenceSets} >
-        <Route path=":id" baseurl="http://1kgenomes.ga4gh.org/" component={ReferenceSet} />
+      <Route path="referencesets" baseurl={baseurl} component={ListReferenceSets} >
+        <Route path=":id" baseurl={baseurl} component={ReferenceSet} />
       </Route>
+      <Route path="variantsets/:variantSetId" baseurl={baseurl} component={VariantSet}>
+        <Route path="variants" baseurl={baseurl} component={SearchVariants} />
+      </Route>
+      <Route path="readgroupsets/:id" baseurl={baseurl} component={ReadGroupSet} />
+      <Route path="featuresets/:id" baseurl={baseurl} component={FeatureSet} />
     </Route>
     <Route path="/about" component={About}/>
  </Router>, document.getElementById('root'));
